requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
fastapi==0.92.0
fonttools==4.38.0
h11==0.14.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.0
numpy==1.24.2
opencv-python==4.7.0.68
packaging==23.0
Pillow==9.4.0
pybase64==1.2.3
pydantic==1.10.5
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
